# LSTM tile size