Changing LSTM GEMM tile size to 8, which gives huge improvement! Also fixing a CMake bug.

I tried skewing with inner loop parallelization as well but it doesn't seem to give any improvement. It probably requires CUDA streams for parallelization (which we don't have). 

The previous benchmarks were on Salike. I switched to Pascal4 (Google cloud) for benchmarks and it's actually pretty close to Nvidia blogpost with this last change:

| Program   | Time  |
|-----------------|-------|
| Old (16 tile)   | 87ms  |
| New (8 tile)    | ~38ms |
| Nvidia Blogpost | ~35ms |

Also Nvidia's code runs slower than they reported (27.8ms). Their run was on Maxwell40, which is an older architecture but maybe it's better than Pascal4.

Finally, I tried to implement a cuDNN benchmark directly from library, but it is extremely tedious and complicated so I stopped halfway through. It's still a better idea to compare it to a library that uses cuDNN as backend.

TODO: We still need to check for correctness.